1. ECTL_EXPANDTABS - Param2 ожидается как поинтер на intptr_t.    x64 плагины надо подправить, если используется ECTL_EXPANDTABS.
2. ошибка в ECTL_SETKEYBAR, когда Param2==-1.

+++ far/editor.cpp
@@ -6129,7 +6129,7 @@ int Editor::EditorControl(int Command, intptr_t Param1, void *Param2)
 			}
 			else
 			{
-				int StringNumber=*(int *)Param2;
+				intptr_t StringNumber=*(intptr_t*)Param2;
 				Edit *CurPtr=GetStringByNumber(StringNumber);
 
 				if (!CurPtr)
